Database access for native image update * Support for Oracle JDBC driver * Support for H2 JDBC driver * Support for mongoDB * Updated Pokemon example to work with all of the above
