--- dbclient/h2/src/main/resources/META-INF/native-image/io.helidon.dbclient/helidon-dbclient-h2/reflect-config.json
@@ -1,10 +1,0 @@
-[
-  {
-    "name": "java.sql.Statement[]",
-    "allDeclaredConstructors": true,
-    "allDeclaredMethods": true,
-    "allPublicMethods": true,
-    "allDeclaredFields": true,
-    "allPublicFields": true
-  }
-]
