+++ dbclient/jdbc/src/main/resources/META-INF/native-image/io.helidon.dbclient/helidon-dbclient-jdbc/reflect-config.json
@@ -1,0 +1,10 @@
+[
+    {
+        "name": "com.zaxxer.hikari.HikariConfig",
+        "allDeclaredConstructors": true,
+        "allDeclaredMethods": true,
+        "allPublicMethods": true,
+        "allDeclaredFields": true,
+        "allPublicFields": true
+    }
+]
