--- dbclient/oracle/src/main/resources/META-INF/native-image/io.helidon.dbclient/helidon-dbclient-oracle/native-image.properties
@@ -1,35 +1,0 @@
-#
-# Copyright (c) 2019 Oracle and/or its affiliates. All rights reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-
-Args= --no-fallback --allow-incomplete-classpath --enable-all-security-services \
-  --initialize-at-build-time=oracle.net.jdbc.nl.mesg.NLSR_en \
-  --initialize-at-build-time=oracle.jdbc.driver.DynamicByteArray \
-  --initialize-at-build-time=oracle.sql.ConverterArchive \
-  --initialize-at-build-time=oracle.sql.converter.CharacterConverterJDBC \
-  --initialize-at-build-time=oracle.sql.converter.CharacterConverter1Byte \
-  --initialize-at-run-time=oracle.jdbc.pool.OracleDataSource \
-  --initialize-at-run-time=oracle.jdbc.driver.OracleDriver \
-  --initialize-at-run-time=oracle.jdbc.driver.OracleTimeoutThreadPerVM \
-  --initialize-at-run-time=oracle.jdbc.driver.BlockSource$ThreadedCachingBlockSource \
-  --initialize-at-run-time=oracle.jdbc.driver.BlockSource$ThreadedCachingBlockSource$BlockReleaser \
-  --initialize-at-run-time=oracle.security.o5logon.O5Logon \
-  --initialize-at-run-time=oracle.net.nt.TimeoutInterruptHandler \
-  --initialize-at-run-time=oracle.jdbc.driver.LogicalConnection \
-  --initialize-at-run-time=oracle.jdbc.driver.NoSupportHAManager \
-  --initialize-at-run-time=oracle.net.nt.Clock \
-  --initialize-at-run-time=oracle.net.nt.TcpMultiplexer \
-  --initialize-at-run-time=oracle.net.nt.TcpMultiplexer \
-  --initialize-at-run-time=oracle.net.nt.TcpMultiplexer$LazyHolder
