+++ examples/dbclient/jdbc/src/main/resources/application.yaml
@@ -86,4 +86,3 @@ db:
     select-for-update: "SELECT * FROM pokemons WHERE name = :name for UPDATE"
     update: "UPDATE pokemons SET type = :type WHERE name = :name"
     delete: "DELETE FROM pokemons WHERE name = ?"
-    # delete-all: "DELETE FROM pokemons"
