+++ examples/dbclient/pokemons/README.md
@@ -26,25 +26,61 @@ Database model contains two tables:
 
 with 1:N relationship between *Types* and *Pokemons*
 
+Examples are given for H2, Oracle, or MySQL databases (note that MySQL is currently not supported for GraalVM native image)
+
+To switch between JDBC drivers:
+
+- Uncomment the appropriate dependency in `pom.xml`
+- Uncomment the configuration section in `application.yaml` and comment out the current one
+
 ## Build
 
+To build a jar file
 ```
-cd <project_root>/examples/dbclient/pokemons
 mvn package
 ```
 
-## Run
+To build a native image (supported only with Oracle, MongoDB, or H2 databases)
+```
+mvn package -Pnative-image
+```
 
-This example requires a MySQL database, start it using docker:
+## Database
+This example can run with any JDBC supported database.
+In the `pom.xml` and `application.yaml` we provide configuration needed for Oracle database, MySQL and H2 database.
+Start your database before running this example.
+
+Example docker commands to start databases in temporary containers: 
+
+Oracle:
+```
+docker run --rm --name xe -p 1521:1521 -p 8888:8080 wnameless/oracle-xe-11g-r2
+```
+For details on an Oracle Docker image, see https://github.com/oracle/docker-images/tree/master/OracleDatabase/SingleInstance
+
+H2:
+```
+docker run --rm --name h2 -p 9092:9082 -p 8082:8082 nemerosa/h2
+```
+For details, see http://www.h2database.com/html/cheatSheet.html
+
+MySQL:
 ```
 docker run --rm --name mysql -p 3306:3306 -e MYSQL_ROOT_PASSWORD=root \
        -e MYSQL_DATABASE=pokemon -e MYSQL_USER=user -e MYSQL_PASSWORD=password  mysql:5.7
 ```
 
+
+## Run
+
 Then run the `io.helidon.examples.dbclient.pokemons.PokemonMain` class:
 ```
-cd <project_root>/examples/dbclient/pokemons
 java -jar target/helidon-examples-dbclient-pokemons.jar
+```
+
+Or run the native image:
+```
+./target/helidon-examples-dbclient-pokemons
 ```
 
 ### Run with MongoDB
@@ -56,7 +92,6 @@ docker run --rm --name mongo -p 27017:27017 mongo
 
 Then run the `io.helidon.examples.dbclient.pokemons.PokemonMain` class with `mongo` argument:
 ```
-cd <project_root>/examples/dbclient/pokemons
 java -jar target/helidon-examples-dbclient-pokemons.jar mongo
 ```
 
@@ -73,8 +108,8 @@ The query operation adds database trace.
 
 `curl` commands:
 
-- `curl http://localhost:8079/db/type | json_pp` - list all pokemon types in the database
 - `curl http://localhost:8079/db/pokemon | json_pp` - list all pokemons in the database
+- `curl http://localhost:8079/db/type | json_pp` - list all pokemon types in the database
 - `curl http://localhost:8079/db/pokemon/2 | json_pp` - get a single pokemon by id
 - `curl http://localhost:8079/db/pokemon/name/Squirtle | json_pp` - get a single pokemon by name
 - `curl -i -X POST -d '{"id":7,"name":"Rattata","idType":1}' http://localhost:8079/db/pokemon` - add a new pokemon Rattata
