--- examples/dbclient/pokemons/src/main/resources/META-INF/native-image/io.helidon.applications/helidon-examples-dbclient-pokemons/native-image.properties
@@ -1,18 +1,0 @@
-#
-# Copyright (c) 2019 Oracle and/or its affiliates. All rights reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-
-Args= --initialize-at-build-time=org.eclipse.yasson \
-   --initialize-at-build-time=org.eclipse.yasson.internal
