+++ examples/dbclient/pokemons/src/main/resources/META-INF/native-image/io.helidon.examples.dbclient/helidon-examples-dbclient-pokemon/reflect-config.json
@@ -1,0 +1,9 @@
+[{
+  "name" : "io.helidon.examples.dbclient.pokemons.Pokemon",
+  "allDeclaredConstructors" : true,
+  "allPublicConstructors" : true,
+  "allDeclaredMethods" : true,
+  "allPublicMethods" : true,
+  "allDeclaredFields" : true,
+  "allPublicFields" : true
+}]
