+++ integrations/db/h2/src/main/resources/META-INF/native-image/io.helidon.integrations.db/h2/native-image.properties
@@ -1,5 +1,5 @@
 #
-# Copyright (c) 2019, 2020 Oracle and/or its affiliates.
+# Copyright (c) 2020 Oracle and/or its affiliates.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -13,6 +13,5 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-
-Args= --initialize-at-build-time=org.slf4j \
-  --initialize-at-run-time=com.zaxxer.hikari.metrics.dropwizard.CodaHaleMetricsTracker
+Args=--initialize-at-run-time=org.h2.store.fs.FileChannelOutputStream \
+     --initialize-at-run-time=org.h2.store.fs.FileNioMemData
