+++ integrations/db/h2/src/main/resources/META-INF/native-image/io.helidon.integrations.db/h2/reflect-config.json
@@ -1,0 +1,16 @@
+[
+  {
+    "name": "java.sql.Statement[]",
+    "allDeclaredConstructors": true,
+    "allDeclaredMethods": true,
+    "allPublicMethods": true,
+    "allDeclaredFields": true,
+    "allPublicFields": true
+  },
+  {
+    "name": "org.h2.engine.Constants",
+    "allDeclaredConstructors": true,
+    "allDeclaredMethods": true,
+    "allDeclaredFields": true
+  }
+]
