+++ integrations/pom.xml
@@ -42,6 +42,7 @@
     <modules>
         <module>cdi</module>
         <module>graal</module>
+        <module>db</module>
     </modules>
 
     <build>
