+++ media/jsonb/src/main/resources/META-INF/native-image/io.helidon.media/helidon-meida-jsonb/native-image.properties
@@ -1,5 +1,5 @@
 #
-# Copyright (c) 2019, 2020 Oracle and/or its affiliates.
+# Copyright (c) 2020 Oracle and/or its affiliates.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -14,5 +14,5 @@
 # limitations under the License.
 #
 
-Args= --initialize-at-build-time=org.slf4j \
-  --initialize-at-run-time=com.zaxxer.hikari.metrics.dropwizard.CodaHaleMetricsTracker
+Args=--initialize-at-build-time=org.eclipse.yasson \
+     --initialize-at-build-time=org.eclipse.yasson.internal
