+++ media/jsonb/src/main/resources/META-INF/native-image/io.helidon.media/helidon-meida-jsonb/resource-config.json
@@ -1,0 +1,7 @@
+{
+    "bundles": [
+        {"name":"yasson-messages"}
+    ]
+}
+
+
