--- media/jsonb/src/main/resources/META-INF/native-image/native-image.properties
@@ -1,21 +1,0 @@
-#
-# Copyright (c) 2019, 2020 Oracle and/or its affiliates.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-
-#
-# Graal native image supports additional configuration from this property file.
-# Yasson requires this to work.
-#
-Args=-H:IncludeResourceBundles=yasson-messages
