+++ tests/integration/native-image/mp-2/pom.xml
@@ -101,17 +101,15 @@
             <scope>runtime</scope>
         </dependency>
         <dependency>
-            <groupId>io.quarkus</groupId>
-            <artifactId>quarkus-jdbc-h2</artifactId>
-            <version>1.2.1.Final</version>
+            <groupId>io.helidon.integrations.db</groupId>
+            <artifactId>h2</artifactId>
             <scope>runtime</scope>
         </dependency>
-        <dependency>
-            <!-- we need this explicit version to work with native image -->
-            <groupId>com.h2database</groupId>
-            <artifactId>h2</artifactId>
-            <version>1.4.197</version>
-        </dependency>
+<!--        <dependency>-->
+<!--            <groupId>io.helidon.integrations.db</groupId>-->
+<!--            <artifactId>ojdbc</artifactId>-->
+<!--            <scope>runtime</scope>-->
+<!--        </dependency>-->
         <!-- Testing -->
         <dependency>
             <groupId>org.junit.jupiter</groupId>
