+++ tests/integration/native-image/mp-2/src/main/resources/META-INF/microprofile-config.properties
@@ -15,9 +15,11 @@
 #
 server.port=8087
 features.print-details=true
-#javax.sql.DataSource.test.dataSource.url=jdbc:h2:tcp://localhost:9092/test;IFEXISTS=FALSE;INIT=CREATE TABLE IF NOT EXISTS GREETING (FIRSTPART VARCHAR NOT NULL, SECONDPART VARCHAR NOT NULL, PRIMARY KEY (FIRSTPART))\\;MERGE INTO GREETING (FIRSTPART, SECONDPART) VALUES ('Jack', 'The Ripper')
 javax.sql.DataSource.test.dataSource.url=jdbc:h2:tcp://localhost:9092/test
-#javax.sql.DataSource.test.dataSource.url=jdbc:h2:mem:test;IFEXISTS=FALSE;INIT=CREATE TABLE IF NOT EXISTS GREETING (FIRSTPART VARCHAR NOT NULL, SECONDPART VARCHAR NOT NULL, PRIMARY KEY (FIRSTPART))
 javax.sql.DataSource.test.dataSource.user=sa
 javax.sql.DataSource.test.dataSource.password=
 javax.sql.DataSource.test.dataSourceClassName=org.h2.jdbcx.JdbcDataSource
+#javax.sql.DataSource.test.dataSource.url=jdbc:oracle:thin:@localhost:1521/XE
+#javax.sql.DataSource.test.dataSource.user=system
+#javax.sql.DataSource.test.dataSource.password=oracle
+#javax.sql.DataSource.test.dataSourceClassName=oracle.jdbc.xa.client.OracleXADataSource
