--- tests/integration/native-image/mp-2/src/main/resources/META-INF/native-image/tests/mp-2/reflect-config.json
@@ -1,14 +1,0 @@
-[
-    {
-        "name": "org.hibernate.dialect.H2Dialect",
-        "allDeclaredConstructors": true,
-        "allDeclaredMethods": true,
-        "allPublicMethods": true
-    },
-    {
-        "name": "org.h2.engine.Constants",
-        "allDeclaredConstructors": true,
-        "allDeclaredMethods": true,
-        "allDeclaredFields": true
-    }
-]
