--- tests/integration/native-image/mp-2/src/main/resources/meta-config.yaml
@@ -1,28 +1,0 @@
-#
-# Copyright (c) 2020 Oracle and/or its affiliates.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-
-sources:
-  - type: "system-properties"
-  - type: "environment-variables"
-  - type: "file"
-    properties:
-      optional: true
-      path: "mp-config.yaml"
-  - type: "classpath"
-    multi-source: true
-    properties:
-      resource: "META-INF/microprofile-config.properties"
-
